Add package for Habitica
A lot of patching was involved here, because the application is riddled
with trackers and analytics APIs and it fetches a bunch of stuff from
external sites.

The trackers are now gone and we also fetch everything locally.

In addition to that, every new user that registers automatically gets a
lifetime subscription, because we also removed all the payment options.

Registrations are only possible when invited.

I've choosen the GPLv3 as the license for this repository because the
upstream code is also licensed under the GPLv3, so we don't run into
incompatibilities.

Furthermore, we currently don't have any assets in this repositery, so
we don't need to make a licensing exception. If this should change one
day, we can still add such an exception.

Signed-off-by: aszlig <aszlig@nix.build>