Split off package expression
Having the NixOS module and the actual package separate makes it a bit
easier to work with it. In addition to that, instead of the getPackage
helper, we now transform the entire set of packages to use their
packageName attribute as their main attribute.

No functional changes, building build-vm.nix leads to the same hash as
before these changes.

Signed-off-by: aszlig <aszlig@nix.build>