Do not load emojis from external site
Our checks for external services and sites were only within the code of
Habitica itself. However, for rendering emojis in markdown, Habitica
uses "habitica-markdown-emoji", which in turn uses their S3 bucket to
fetch those images.

My first try was to write a small fetcher, which fetches all the files
from the bucket and puts it into a fixed-output derivation.

However, this isn't a very good solutions as the contents of these files
are subject to change.

After fetching a few images and I visually compared them to the Emoji
Cheat Sheet[1] and found that they were the same, so I'm now using the
files from their GitHub repository.

[1]: https://www.webpagefx.com/tools/emoji-cheat-sheet/

Signed-off-by: aszlig <aszlig@nix.build>