Disable caching in nginx config
The problem here is that all our static files are in the Nix store, so
their timestamps are UNIX+1 essentially. This in turn messes up caching,
because if if_modified_since is enabled (which it is by default) the
server checks the timestamp supplied by the client and if it matches it
returns HTTP 304, which it will return even on changed files because the
time stamp is still the same.

Same problem arises with the ETag header, because the way nginx
calculates the ETag is using the time stamp and the file size.

Signed-off-by: aszlig <aszlig@nix.build>