Add helper for generating patch files
This is useful whenever we have a new version of Habitica coming down
the pipes and we want to add all the rebased patches to the patches/
subdirectory.

So instead of doing this manually, all we need to do is running this
script which being in a git clone of Habitica with all the patches.

Signed-off-by: aszlig <aszlig@nix.build>