Replace all hardcoded URIs
There are a lot of places throughout the source code which either
hardcode domains or email addresses and we're also guilty of repeating
that with our "one-admin-mailaddr.patch".

My email address thus is no longer hardcoded in that patch and we're
using sed now to replace all occurences of URIs with the ones either
provided in ADMIN_EMAIL or BASE_URL.

This now makes it possible to use this for instances other than mine.

Signed-off-by: aszlig <aszlig@nix.build>