Split dependencies from main derivation
So far we have provided Habitica itself to the list of dependencies and
also built Habitica using the node-env provided by node2nix.

However, this has several drawbacks, ranging from extensibility to
unnecessary overhead during build. Now the main build takes around 40%
less time, which is still not ideal but at least we die less while
waiting.

Unfortunately we needed to introduce a few workarounds which are
revolving around handling of NODE_PATH. It seems that different node
modules such as babel and webpack try to be clever by implementing their
own module resolving mechanism that entirely bypass NODE_PATH.

Another change is that we now prepare the source tree in a dedicated Nix
expression file (source.nix) which also includes all of our patches.

Having all patches in one place also has the benefit that we can easily
apply them to a fresh Git clone of Habitica in the right order.

The only drawback to what we had before is that we now have to manually
bump the version number of Habitica, which was done by the old
update.sh. This also has an advantage however, because if we just want
to update dependencies without updating Habitica itself, we can just do
that without the need to change update.sh.

Our new updater is now called update-deps.py, so the name also reflects
what it is doing.

Signed-off-by: aszlig <aszlig@nix.build>