Fix loading of client for non-root paths
If a path like /party is requested the server sends the client's
index.html to the browser and it's doing so by referencing the path to
its index.html.

However, the client's index.html is in an entirely different store path
and express doesn't allow absolute paths.

So what we do instead is just embed the client's index.html into a JS
string in libs/client.js and return this instead.

Another approach would be a redirect, but that wouldn't get the user to
the right sub-page of Habitica but to the initial main page instead.

Signed-off-by: aszlig <aszlig@nix.build>