Fix dependency updater
With the split of of the derivations, the updater no longer works
because it's using the "src" attribute of habitica.nix. The reason this
has worked was because there was only a single derivation in
habitica.nix.

So we now just propagate the "src" attribute as "source" in habitica.nix
and use that instead.

Signed-off-by: aszlig <aszlig@nix.build>