Fix spell/skill casting in client
The reason why this doesn't work is a circular import which was
happening in spells.js, which was importing statsComputed.js. That in
turn imported content/index.js which again imported spells.js.

So the result of the statsComputed.js from spells.js has caused
statsComputed to be undefined, which is a very weird way of handling
those errors. But well... JavaScript, what should I say?

Signed-off-by: aszlig <aszlig@nix.build>