Fix import of moment-recur
So far this has been imported like this:

import moment from 'moment';
import 'moment-recur';

This is also the way the moment-recur is supposed to be imported.

Fortunately, moment-recur propagates the modified moment module, so we
can solve this simply by just importing moment-recur *instead* of moment
to get a modified moment with moment-recur.

Signed-off-by: aszlig <aszlig@nix.build>