Allow to buy an unlimited amount of gems
By default the limit of the gems a user can buy is based on his/her
subscription plan. However, on our instance we don't have payments, so
it doesn't make sense to limit this.

This removes the number of the top-right edge of the gem item and
doesn't impose the limit anymore.

Signed-off-by: aszlig <aszlig@nix.build>