Don't put secrets inside the Nix store
This removes SESSION_SECRET, SESSION_SECRET_KEY and SESSION_SECRET_IV
from the configuration file and passes them via the environment instead.

The environment file that is used for that is generated once for the
first start up of Habitica and it's populated by random values.

After that, the environment file is then used for the main Habitica
service using the EnvironmentFile unit option.

Signed-off-by: aszlig <aszlig@nix.build>