Remove Instagram link in news.js
There is a canary for Instagram and I previously added an exception for
news.js. However, if people want to follow the Instagram account of
HabitRPG Inc. then they can still go to the official site.

Apart from that there is advertisement for their Twitter, Tumblr and
Facebook page, which I also removed.

Signed-off-by: aszlig <aszlig@nix.build>