Remove staff and tiers section in tavern
The staff links were broken anyway, because those users do not exist on
a non-official instance.

Tiers on the other hand would make sense only if our goal is to make a
single instance as well, so we could honor contributors by giving them
shiny colors. For inofficial self-hosted instances however I don't think
this is useful, so let's remove it. We can still bring it back if people
want it.

Signed-off-by: aszlig <aszlig@nix.build>