Remove press kit, merch and group plans
Those are mostly advertising and most of that information also doesn't
apply to our standalone instance.

In addition to that, I had to break out groupPlans.vue from the
remove-external-services.patch, because we now delete the file entirely
within our Nix build.

We new also delete the presskit, merch and emails directories in the
static directory. The reason for deleting the emails directory is
because those assets are only used by HabitRPG's MailChimp instance and
they're really not relevant to us.

Signed-off-by: aszlig <aszlig@nix.build>