patches: Fix all eslint errors
I've prepared checking with eslint and along the way it found a bunch of
errors, most of them due to unused variables or imports.

Some of them however were real errors I made, for example missing
semicolons at the end of an expression, using !! instead of Boolean(),
comparing strings with == and using double quotes instead of single
quotes for string constants.

Signed-off-by: aszlig <aszlig@nix.build>