Check source with ESLint after applying patches
We don't want to introduce new errors to the game, so let's run ESLint
in the checkPhase.

I've also made sure that we avoid adding NODE_PATH and eslint to
nativeBuildInputs, because we don't want that crap in the build
environment if someone uses overrideAttrs to set doCheck to false.

Signed-off-by: aszlig <aszlig@nix.build>