Add additional function canaries and helper tool
We want the builder to fail whenever there is a function call using a
different amount of arguments to what we have already patched. This is
to prevent upstream changes from interfering with our changes.

The helper tool is to find those canaries in our patches and it spits
out the location of these plus the suggested regexes to use.

Signed-off-by: aszlig <aszlig@nix.build>