Add the ability to run tests
I had to patch the test runners so that they properly run in our
environment, because they relied on node_modules being available and
also had some limit on the output buffer.

Apart from that, all tests that are specific to external services are
now removed as well, but we don't execlude the test directory from the
canary checks, because there are no external services that could
magically "leak in" and if a test refers to such a service it will fail
anyway.

Tests are run using mkCommonBuild and their respective gulp target and
those tests needing database connectivity also get an ad-hoc
"ephemeralForTest" MongoDB instance.

Signed-off-by: aszlig <aszlig@nix.build>