Generate nice HTML/CSS test reports
We also link them as Hydra build products and prevent the actual build
to fail, touching nix-support/failed to signal Hydra that the test has
actually failed. That way we still get the HTML report.

The patch I added uses mocha-simple-html-reporter to generate the report
and mocha-multi-reporters so that we get the report *and* the output to
the terminal if we don't build on Hydra.

Signed-off-by: aszlig <aszlig@nix.build>