README: Add idea about e2e encryption
For more privacy this is something that might be very useful, especially
because even with small deployments it doesn't make a lot of sense to
play the game only by yourself. So in a way, you still have a similar
problem as with the official instance, but instead of HabitRPG Inc.
being able to read the task data the admin of that instance can read
them.

Signed-off-by: aszlig <aszlig@nix.build>