tests: Add client:e2e test
We have disabled launching of Selenium by Nightwatch in our
nightwatch.patch before and we launch the Selenium Server in the test as
a systemd service.

The service needs to run as non-root, because otherwise Chromium will
complain that --no-sandbox can only be used if it's running with normal
user privileges.

I didn't want to dig further into how to properly set the path to
Chromium in chromedriver, because the client:e2e test is somewhat
useless and there is only one single test case.

Signed-off-by: aszlig <aszlig@nix.build>