Update to version 4.35.1
Upstream changes:

  * New Armoire and Backgrounds
  * Update antidote display and functionality
  * Enable Shiny Seeds

There was also an April Foolery update starting at version 4.34.2 and
which got reverted in version 4.35.1. This has scaled all the pets and
mounts to 33% of their original size.

I'm so glad we skipped this :-)

We now no longer need the patch on news.js, because it no longer
contains external links.

Signed-off-by: aszlig <aszlig@nix.build>