node-overrides: Fix version match of versionOlder
When removing the assertion on the version attribute in commit
60fef83addeb7eb5c9b8520d0e951636052f2c16, I din't change the attribute
refering to libsass (which was 'drv' within the override).

However, there is also a 'drv' from the outer scope which was shadowing
the inner 'drv' attribute, but with the override gone, that attribute
now refers to node-sass, which currently has a version of "4.8.3", so
the version is never older that 3.5.0.

Signed-off-by: aszlig <aszlig@nix.build>