Remove censorship in chat messages
Every chat message in public chat gets checked against a list of bad
words and slurs and prevents them from being used (and also sends a
notification to the mods).

We have a standalone instance that everybody can set up by themselves
and it's invite-only by default, so if there are people misbehaving it's
their own responsibility which people they invite and whether they
tolerate such words being used or not.

Signed-off-by: aszlig <aszlig@nix.build>