source: Clean up canary exceptions
Ideally we don't want to have any exceptions, but for now, let's get rid
of most of them, because they're either already patched out or are now
patched out.

One of them is that now the functionality for cancelling a user's
subscription (and its tests) doesn't exist anymore, because we don't
want anyone to accidentally cancel their subscription (it's not so easy
to get it back again).

In the long term of course, we want to get rid of all the code related
to subscriptions.

Another thing I did while rebasing the patches is to reorder
remove-external-services.patch to be the first patch to be applied. This
makes it easier to fixup up patches that are ordered afterwards.

Signed-off-by: aszlig <aszlig@nix.build>