build-vm: Allow to easily open connections to VM
So far there was only one main SSH connection to the guest, which also
controls whether to shut down the VM. However, sometimes it's very
useful to be able to establish additional connections for example to
transfer files.

This is now possible with a new command line flag "--connect", which
basically does the same as the main connection but allows to pass flags
to the SSH command.

For example something like "./result --connect tar c foo | tar x" will
copy over "foo" from the root user's home directory.

Signed-off-by: aszlig <aszlig@nix.build>