Add new insecureDB option
This is solely used for development and testing and as its name implies
shouldn't be used in production, especially on a multi-user system.

We're going to use this however to export/import test fixtures, because
mongodump and mongorestore do not have a way to connect to UNIX sockets
yet (or at least I haven't found a way).

Signed-off-by: aszlig <aszlig@nix.build>