source: Add patch adding substvar for migrations
We're going to have a new file that lists all the migrations from the
start of this fork to the current version, so we can run all of the
migrations until we have a database schema that hopefully[TM] matches a
fresh install.

The upstream repository has a migration runner for that, which is just a
small wrapper around a specific migration file that is executed after
the database connection has been set up. We do this as a patch because
we want the build to fail if some of the semantics of the migration
runner changes in some future update.

Signed-off-by: aszlig <aszlig@nix.build>