Patch MongoDB to use systemd startup notifications
So far the reason why Habitica didn't run into connection errors is
because Habitica is socket activated and thus doesn't start directly
during the boot process but later when requested.

Also, we're going to integrate database migrations, so we need to be
sure that the database system has been started and is accepting
connections. While this could have been done by using a wait/retry loop
in a shell script, I didn't do it that way because I consider this a
workaround.

What we have now instead is a very tiny patch against MongoDB which uses
sd_notify() to signal startup completion.

Signed-off-by: aszlig <aszlig@nix.build>