Switch to use systemd startup notifications
This is really useful for testing but also makes sure that if habitica
is ordered before a certain target we can be sure that it's started up
and accepting connections.

The implementation is done by using the sd-notify node binding, which in
contrast to the systemd-notify NPM library is actually a native binding.

Signed-off-by: aszlig <aszlig@nix.build>