Make habitica.service part of multi-user.target
Especially when running the development VM (via build-vm.nix) it's very
annoying when we have to wait for the machine startup *and* the startup
of Habitica after we're initiating the first request.

Now we only need to wait for the machine startup, which also comes in
handy for NixOS VM tests.

Signed-off-by: aszlig <aszlig@nix.build>