Allow to create periodic database backups
I mainly needed this in order to implement database migrations, so that
we can make a backup of the current database prior to running the
migration scripts.

However, it turned out that the backup mechanism is something that's
generally good to have, so we're now exposing this mechanism via the new
backupInterval option.

The hardest part here was to implement the NixOS VM test, because we
needed to do a few API calls in order to create data that's guaranteed
to be restored after habitica-db-restore. My first attempt was just
creating an initial backup, creating a user, restore the backup and see
the user gone. However this won't work, because mongorestore doesn't
drop the whole database.

While this could be solved by dropping the whole database prior to
restoring, I didn't use that route because a failed restore would still
give us no users and the test then would succeed even though things went
wrong.

Signed-off-by: aszlig <aszlig@nix.build>