tests: Add DB fixture for upcoming migration test
I always try to implement tests first, so let's start with the basic
data we need for the migration test, which is a snapshot of Habitica's
database at version 4.38.0.

The Nix expression to create the fixture is using specific revisions of
nixpkgs and this very repository so that it's reproducible every time.

One exception that might not be reproducible is the habitpy.nix file,
because it will be used for the actual test as well and if it's changed
during further iterations it might not be backwards-compatible. In the
worst case however, we can still duplicate the expression later.

For the test fixture we just create a new user and post a chat message,
because that's what has changed in version 4.40.0, if we want to test
other things, we now can quickly regenerate the fixture.

Signed-off-by: aszlig <aszlig@nix.build>