habitica.nix: Add migration runner
This is going to be used in the systemd service that's going to be
started between the database and Habitica itself.

Basically we're using the same migration runner from upstream but
instead of starting a full server, we just require the relevant files
needed for the migrations.

The actual migration files are read from migrations.nix and those are
numbered with an index of 1, because the state in Habitica 4.38.0 is
schema version 0 and we want to upgrade _to_ version 1, so beginning
with 1 is IMHO more sensible.

For transpiling the migration files we use a custom gulp file, so we
don't need to add babel-cli to our node dependencies.

Note that the reason for using a list of attribute sets for
migrations.nix instead of a list of strings is because we're going to
directly add test scripts to the attribute set so that the tests and the
actual migrations are close together.

Signed-off-by: aszlig <aszlig@nix.build>