Implement support for database upgrades
We now use our previously written migrator to run all of the upgrades
specified in migrations.nix. Prior to doing the migrations, we create a
database backup so that if such an update fails, we can recover. The
recovery currently has to be done manually, but we might automate this
in some future version.

The version of the schema currently is written into
/var/lib/habitica/db-version and it's checked by a small helper tool,
which then performs the updates if the version doesn't match.

Signed-off-by: aszlig <aszlig@nix.build>