Add a test for migrations
This is what I have used to implement the migration support until the
test succeeds.

Every migration (currently there is just one) specified in
migrations.nix now also has a testScript snippet, which is executed in
the context of a Python unittest.TestCase function with self being made
available to the local scope.

The advantage of this is that there is no occurence of "self" in the
testScript, so it won't be confusing in terms of people wondering where
that "self" came from. It also allows us to limit the variables
available within that scope.

Signed-off-by: aszlig <aszlig@nix.build>