tests/backup: Don't fast-forward time
While this is working perfectly on NixOS master, where we have systemd
version 238, it doesn't work for NixOS 18.03, which uses systemd version
237.

The reason for that is systemd/systemd#8231, where the fix has been
merged in systemd/systemd@9ea9faff7866ac7aaa2dc529fb92c5cdd643fc04.

Version 237 of systemd doesn't have that change yet, so in order to also
simplify things a bit, let's actually trigger the backup service
manually instead and remove all that cruft for waiting until the backup
is done as the manual trigger is synchronous.

Unfortunately though, this removes test coverage for backupInterval, but
we can still revert this change once we switch to NixOS 18.09.

Signed-off-by: aszlig <aszlig@nix.build>