tests/migrations: Ensure migrations are run once
We don't want migrations to run twice or even more often, so let's
actually test that by restructuring the way we populate the database
with old data.

Instead of using a systemd service, we do it within the testScript now,
which has the advantage that we can simply restart the machine without
the need to change system configuration.

In the end we restart the machine twice: The first time after populating
the database with old data and the second time to make sure that
migrations didn't run during the third boot.

Ensuring that is done by checking the journal whether "Running
migration" shows up. The string for that now also resides in
docinfo.nix, so that whenever the string changes we still are able to
match the journal entry if the implementation changes and we forget to
change the test.

This also gets rid of using insecureDB, because we now have
habitica-db-tools to connect directly via UNIX socket.

Signed-off-by: aszlig <aszlig@nix.build>