Add patch to fix HabitRPG/habitica#10333
The issue was that whenever players were casting a spell using the
client (not when using the API directly), it was casted twice instead
and the quantity wasn't checked on the server side, so a player who
didn't have enough mana for casting the spell a second time, the MP
would be negative.

Thanks to @TheHollidayInn for pushing a fix in HabitRPG/habitica#10337,
which essentially is a revert of HabitRPG/habitica@8d25a5d140958faa790e.

I've tested this shortly with build-vm.nix and it indeed fixes the
problem.

Signed-off-by: aszlig <aszlig@nix.build>