Fix translation string on home page
Until this change the start page contained:

  Motivate yourself to achieve your goals.

  Error processing the string "timeToGetThingsDone".
  Please see Help > Report a Bug.

The reason is that a new translation variable got introduced in the
meantime after our invite-only.patch was made. So we need to add the
userCountInMillions. We don't have that amount of users for standalone
instances, so in the long term we might want to do a complete overhaul
of the home page.

Signed-off-by: aszlig <aszlig@nix.build>