Strip markdown from claimed task messages in group
This has been annoying me for quite a while, because whenever a task
contains Markdown the system message will print the Markdown as-is.

It's even more annoying if the task contains backticks, because that
will mess up the system messages even further as the whole message is
already wrapped inside backticks, so the inner backticks will simply
close the code block.

Signed-off-by: aszlig <aszlig@nix.build>