Update to version 4.42.3
Version 4.42.3:

  * Disable seasonal potions

Version 4.42.2:

  * Downgrade Bootstrap to restore tag checkboxes

Version 4.42.1:

  * Fix notifications in user pre save hook
  * Fix level up webhook
  * Improve sorting Party header
  * Refactor payments
  * Fix columns when translation is too long
  * Fix learn more in user dropdown
  * Add responsive margins to pets and mounts
  * Fix header stuck naming
  * Fix placeholder and the message row
  * Do not load inbox in tasks routes
  * Do not load inbox in some user routes
  * Move developer-only strings to api/common messages

I needed to address the move of the developer-only strings, because we
now have an additional directory called "errors" in common/, so we need
to handle it the same way as locales because our directory structure is
different from upstream.

Another thing which got introduced in version 4.42.2 was the downgrade
of Bootstrap, which was still using caret notation in package.json for
the version. This essentially does not downgrade anything because it
still selects Bootstrap 4.1.1 instead of 4.1.0 and we're not using
package locks.

Signed-off-by: aszlig <aszlig@nix.build>