Fix error due to empty media src attribute
The added patch fixes the following error occuring in Firefox:

Invalid URI. Load of media resource  failed.
All candidate resources failed to load. Media load paused.

The reason for this error is that Habitica is setting the src attributes
to an empty string and filling it once an audio file is going to be
played.

Signed-off-by: aszlig <aszlig@nix.build>