Update to version 4.44.3
Upstream changes relevant to us:

  * Use Promise array when gifting gems
  * Remove parallel calls to save()
  * Remove Cuddle Buddies Bundle from featured items

Rebasing the patches went without conflicts, so this update is a rather
quiet one. However, it also disables a range of tests that would work
for us because we use libfaketime. I've fixed the tests already and the
changes are upstream, so we might want to cherry-pick them.

Signed-off-by: aszlig <aszlig@nix.build>