Update to version 4.48.0
API changes without GitHub mentions:

  * Feature: Include timestamp of flagged chat message sent to Slack
             integration
  * Feature: Initial support for new API version 4
  * Feature: Inbox messages GET route for API v4
  * Content: Subscriber mystery items for June 2018
  * Fix: Address async issue that could cause crons to run twice in a
         very short period
  * Chore: Locale files updated
  * Chore: Bailey news announcements

Client changes without GitHub mentions:

  * Fix: Obtain user's party status (no party, solo party, party
         w/members) to decide what to display when clicking party view
         buttons
  * Fix: Better margins on boss monster data

The release notes including mentions can be found at:

https://github.com/HabitRPG/habitica/releases/tag/v4.48.0

While rebasing the patches I resolved a few conflicts regarding the
work-in-progress version 4 of the API, which moved a few files around.

This also made it necessary to rename a few paths that are in our list
of files to kill during source code preparation.

In addition to that I also added the new integration test for APIv4.

Signed-off-by: aszlig <aszlig@nix.build>