tests/create-fixture: Hardcode initial start time
For the upcoming migrations, the format of the task histories will
change, so whenever we create the fixture with multiple task scores we
want to make sure that whenever a task is scored again that it will stay
on the same day.

The format will merge all the scores for a specific task history based
on the day it was made, so we ensure that we stay on the same day by
setting the machine's date to 12:00 on January 1st of 2018.

Signed-off-by: aszlig <aszlig@nix.build>