tests/create-fixture: Add time warp functionality
In order to score tasks on different days we need to have a way to
change the current time of the VM, for which we now have a handy
timewarp() function that will pass the amount of time to fast-forward to
a socket which in turn sets the time of the machine accordingly.

This functionality isn't used so far but will be used in upcoming
fixture creation changes.

Signed-off-by: aszlig <aszlig@nix.build>