update-patches: Find unannotated tags as well
When figuring out the base revision, we so far only used annotated tags
but version v4.50.4 is unannotated. For git describe this is usually to
avoid using tags that are unrelated to the version, but in our case our
patch series is always based on a version and the latest tag won't be in
the middle of our patch series.

Signed-off-by: aszlig <aszlig@nix.build>