Add very rudimentary shell.nix
At the moment this only sets up a small sandbox (FS namespacing only)
and mounts the source tree given by sourceTree into the FS namespaced
chroot.

The goal here is to have a simple way to run nix-shell on a detached
source tree where we can make changes and immediately see the results.

Our sandboxing makes sure that the source tree is read only and we won't
end up with crap cluttering the directory with random files.

Right now however, this still has a few issues.

First of all, we do have a few files that we kill off in the source
tree, but instead of using patches we do it in the source code
preparation phase of the build to avoid having ultra-large patches. This
however isn't working very well if we have a detached git tree which
still contains these files, so we need to find a way how to make this
available to the sandbox without the need to alter the source code.

Second, we currently do not get a real shell but it just starts up
serving Habitica locally, so people used to nix-shell might be confused
about that.

Signed-off-by: aszlig <aszlig@nix.build>