Add a helper to remove killed files
With filesToKill being exposed in source.nix, we now have a small helper
utility which can quickly remove all the files from the source tree that
we do not want to have.

Running rm -r on these files however is a bit scary, so we might want to
switch to a better solution in the future.

Signed-off-by: aszlig <aszlig@nix.build>