Update to version 4.51.0
The only upstream change relevant for us are the new Armoire items and
avatar backgrounds for July 2018.

Other than that, it's a very quiet update with no conflicts but with a
rename of "Dockerfile-Production" to "Dockerfile", so I had to adjust
the canary exclusion accordingly.

With this update we can now drop the fix-version.patch, which obviously
has been fixed upstream.

Signed-off-by: aszlig <aszlig@nix.build>