member-details: Only add 1px margin when condensed
Adding a 1 pixel right margin to all nodes with the member-stats class
affects the display when you click on the avatar as well, which will
then break the layout because of that additional margin.

So let's just do this for the condensed version.

This has been introduced by HabitRPG/habitica#10407 and I've addressed
this upstream at HabitRPG/habitica#10504.

Signed-off-by: aszlig <aszlig@nix.build>