habitica.service: Restrict file system access
While we already run it within a network namespace, the sandboxing is
still too lax for my tastes. If an attacker can compromise the service
they have access to the full file system (although with user
privileges).

So what we now do is gathering all the paths for the habitica.server
runtime closure and set up a file system namespace which has
bind-mounted the relevant store paths plus the sockets for the mailer
and the database system.

There is a goof however, namely that if a store path is a symlink it
currently is ignored. For example libGL is a symlink to libglvnd on
current NixOS Unstable.

My first attempt at solving this was to use ExecStartPre for creating
the symlinks. However systemd sets up the namespacing + bind mounts for
*every* execution, so if we do make the symlinks during ExecStartPre
they won't be available for the main process anymore because the tmpfs
for the ExecStartPre instances have already been dissolved.

I dug through the systemd source code a bit and found no suitable way to
do that, but given that libGL is not needed for the Habitica server to
run, I'm simply skipping all symlinks for now until I found a way to do
this.

Signed-off-by: aszlig <aszlig@nix.build>