Switch to and introduce an external mailer daemon
So far the Habitica server has directly invoked the sendmail binary,
which was not only very ugly having the mail templates inline but it
also has some security implications.

The main implication here is that if we want to sandbox the Habitica
server we'd need to give it access to a setuid binary, which is not a
very good idea as it increases the attack surface tremendously.

With the new mailer daemon, all we need to have is a socket which
Habitica communicates with. The socket also is provided by a systemd
socket, so we have all mail handling isolated within the mailer daemon.

In the future we can also implement stuff like opt-in directly within
the mailer, so whenever a user registers with a email address of
somebody else we can first check whether the address is really valid.

At the moment, the mailer pretty much replicates what we had so far,
supporting only the welcome, invite-friend and unknown template. But we
now also support password reset emails.

Another advantage of the mailer daemon is that we now have a lightweight
way to run tests on the templates, without the need to have Habitica
running.

Our mailer daemon roughly tries to mimick the same protocol which
Habitica uses in the first place, so we can keep our new "mailer-daemon"
patch to be minimally invasive.

I have also modified the mailer VM test to include the password-reset
mail.

Signed-off-by: aszlig <aszlig@nix.build>