tests: Add a very basic test for emails
We're going to replace the email functionality we had so far with a
mailer daemon which replicates MailChimp's API. That way we need less
patching of Habitica's code.

The test here is only in preparation for that and it currently only
makes sure that we get *any* email (in this case the party invitation
mail), but no verification of the email is done yet.

Signed-off-by: aszlig <aszlig@nix.build>