mailer: Fix handling of personalVariables
Turns out (yes, Captain Obvious incoming) that the personalVariables
object is a mapping between the recipients and variables specific to
them, as the name implies.

So we now use a map from the recipient address to the personalVariables
set for the recipient, so that not everybody gets the same unsubscribe
link (which will unsubscribe the first player).

We now render the template for each recipient instead of just once, so I
also generalized the tests a bit so we have a bit less redundant code,
even though there is still more to be DRYed.

Signed-off-by: aszlig <aszlig@nix.build>