Don't use TemporaryFileSystem with systemd 237
Unfortunately, systemd 237 doesn't support the TemporaryFileSystem unit
option, silently ignores it and then fail when trying to start the
service.

So as a workaround for systemd 237, we now have a /run/habitica-chroot
directory mounted as a tmpfs, which is then bind-mounted into the
chroot.

The reason why we need to be backwards-compatible to systemd 237 is
because it's the version used in the latest NixOS stable (18.03), so we
need to wait for 18.09 until we can drop this workaround.

Signed-off-by: aszlig <aszlig@nix.build>