Improve dependencies for chroot tmpfs
Let's remove the automount, because it really isn't needed and we only
need a mount unit that binds to the main habitica.service, so that
whenever the service is stopped, the tmpfs mount gets cleaned up as
well.

I also removed the rbind option within the main service, because we
really don't care about nested mountpoints as we set up our own mount
namespace anyway.

Signed-off-by: aszlig <aszlig@nix.build>