Make the mailer daemon purely socket-activated
The mailer daemon is only useful for habitica.service and thus we
shouldn't pull it into multi-user.target.

Apart from that, let's make *only* the mailer daemon's socket a
requirement of habitica.service, so that it's always guaranteed the
socket exists prior to starting up Habitica and if the habitica.service
is disabled we really don't want the socket.

Signed-off-by: aszlig <aszlig@nix.build>