Update to version 4.55.0
This contains only locale updates (which include the missing
translations for the Mystery Set 2018-07) and the fix for the english
locale we have already cherry-picked.

Apart from that there is only a news update, which doesn't affect us.

So this update is mainly relevent for us in that we can now drop the
cherry-picked commit :-)

Signed-off-by: aszlig <aszlig@nix.build>