Add image proxy to fetch images in markdown
So far everything is pretty well contained and the browser on the client
side doesn't send out requests to external sites.

However, this isn't the case for images used in markdown, for example in
tasks or chat. This not only is a privacy issue but also makes browsers
indicate a mixed-content warning if the image is fetched from an HTTP
URL.

For the image proxy implementation, I first tried to evaluate a service
written in Go from https://github.com/willnorris/imageproxy. There are
also image proxies out there written in JavaScript, but I didn't feel
very comfortable running these because sandboxing them requires a whole
chain of runtime dependencies that need to be accessible within the
chroot.

So the major threats we need to be aware of are:

  * DNS rebind or probing internal ports in general
  * Compromise of the image proxy itself
  * DoSing external sites through the proxy
  * DoSing the proxy itself
  * Serving malicious images (containing rendering or parser exploits)
    via the proxy

After trying to fix the aforementioned image proxy, I decided to scrap
it alltogether and went to write my own implementation.

While thinking about what the implementation would need I went for a
minimal approach without caching, because that's something nginx can do
better. After defining the functionality I rechecked the web and came
across go-camo (https://github.com/cactus/go-camo), which came pretty
close to what I wanted to have and only minor patches were needed.

Of course I also took a look at Camo (the node.js implementation) to see
whether it would also be eligible, but it completely lacked protection
against DNS rebind attacks.

The Go counterpart on the other hand does have protection against this
and also has quite sane limits per default. For the most points above I
have sandboxed the service to make it harder to compromise the whole
system should the proxy be compromised. The only point which is not
addressed is the last point, for which the unfortunately doesn't have
any protection implemented.

At some point we might want to implement this by ourselves, but it's
only a minor concern as nowadays mainstream browsers usually employ
sandboxing and are regularily updated.

In order to only allow Habitica users to send requests to the proxy,
I've implemented a small patch, which checks the signature of the
already existing Habitica session and only allows requests with valid
session cookies.

Given that currently the E2E client tests from upstream Habitica are
unused, I added a test which checks whether images in Markdown are
correctly rewritten. In addition to that I also added a NixOS VM test,
which tests the proxy and the caching (the latter is done via nginx).

Signed-off-by: aszlig <aszlig@nix.build>